+++ CDS/scripts/js_utils/add-asset-local.sh
