+++ CDS/scripts/js_utils/hmac.js
