+++ CDS/scripts/js_utils/test/TestHmac.js
