+++ CDS/scripts/js_utils/youtube/youtube-upload-lib.js
